move W_Cell into its own file and document it
