+++ spy/vm/modframe.py
@@ -4,7 +4,8 @@ from spy.analyze.scope import ScopeAnalyzer
 from spy.analyze.symtable import SymTable
 from spy.errors import SPyError
 from spy.vm.vm import SPyVM
-from spy.vm.module import W_Module, W_Cell
+from spy.vm.module import W_Module
+from spy.vm.cell import W_Cell
 from spy.vm.object import W_Object
 from spy.vm.function import W_ASTFunc
 from spy.vm.astframe import AbstractFrame
